workaround an issue with the nokogiri gem and old rubygem versions (< 1.7)
This is in wait of a proper fix in autoproj. On these old ubuntu
versions, we install both the gem and the os package, so that the
OS package can be used when the gem fails to load (it does install
fine). The proper fix will be to allow installing gems only on
specific OS versions in autoproj.